refactor: User 유효성 검증 패턴 중복 제거
