--- src/main/java/techcourse/myblog/domain/IllegalUserException.java
@@ -1,7 +1,0 @@
-package techcourse.myblog.domain;
-
-public class IllegalUserException extends RuntimeException {
-    public IllegalUserException(String msg) {
-        super(msg);
-    }
-}
