+++ src/main/java/techcourse/myblog/domain/InvalidUserException.java
@@ -1,0 +1,7 @@
+package techcourse.myblog.domain;
+
+public class InvalidUserException extends RuntimeException {
+    public InvalidUserException(String msg) {
+        super(msg);
+    }
+}
