+++ src/main/java/techcourse/myblog/domain/User.java
@@ -3,6 +3,7 @@ package techcourse.myblog.domain;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.DynamicUpdate;
 import org.hibernate.annotations.UpdateTimestamp;
+import techcourse.myblog.validation.UserPattern;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
@@ -30,13 +31,6 @@ public class User {
     @UpdateTimestamp
     private LocalDateTime updateTimeAt;
 
-    @Transient
-    private static final String namePattern = "^[^ \\-!@#$%^&*(),.?\\\":{}|<>0-9]{2,10}$";
-    @Transient
-    private static final String emailPattern = "^[0-9a-zA-Z]([-_.]?[0-9a-zA-Z])*@[0-9a-zA-Z]([-_.]?[0-9a-zA-Z])*.[a-zA-Z]{2,3}$";
-    @Transient
-    private static final String passwordPattern = "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[$@$!%*?&])[A-Za-z\\d$@$!%*?&]{8,}$";
-
     private User() {
     }
 
@@ -50,20 +44,20 @@ public class User {
     }
 
     private void validateName(String name) {
-        if (isEmpty(name) || !Pattern.matches(namePattern, name)) {
-            throw new IllegalUserException("이름은 2~10자, 숫자나 특수문자가 포함될 수 없습니다.");
+        if (isEmpty(name) || !Pattern.matches(UserPattern.NAME_PATTERN, name)) {
+            throw new InvalidUserException(UserPattern.NAME_CONSTRAINT_MESSAGE);
         }
     }
 
     private void validateEmail(String email) {
-        if (isEmpty(email) || !Pattern.matches(emailPattern, email)) {
-            throw new IllegalUserException("이메일 양식을 지켜주세요.");
+        if (isEmpty(email) || !Pattern.matches(UserPattern.EMAIL_PATTERN, email)) {
+            throw new InvalidUserException(UserPattern.EMAIL_CONSTRAINT_MESSAGE);
         }
     }
 
     private void validatePassword(String password) {
-        if (isEmpty(password) || !Pattern.matches(passwordPattern, password)) {
-            throw new IllegalUserException("비밀번호는 8자 이상, 소문자, 대문자, 숫자, 특수문자의 조합으로 입력하세요.");
+        if (isEmpty(password) || !Pattern.matches(UserPattern.PASSWORD_PATTERN, password)) {
+            throw new InvalidUserException(UserPattern.PASSWORD_CONSTRAINT_MESSAGE);
         }
     }
