+++ src/main/java/techcourse/myblog/validation/UserInfo.java
@@ -1,0 +1,4 @@
+package techcourse.myblog.validation;
+
+public interface UserInfo {
+}
